Improve type annotations and type correctness
This adds a number of missing type annotations to the public API surface.

It also fixes the signature of several methods which allow a `follow_symlinks`
parameter in Python 3.13+. This parameter is allowed in `S3Path`, but ignored.

Unfortunately, these changes are not enough for MyPy to run without errors.